update wechat payment

+++ app/models/payment.go
@@ -10,7 +10,7 @@ type Payment struct {
 	OrderNo          string     `json:"order_no" bson:"order_no"`
 	Platform         string     `json:"platform" name:"支付平台:微信/支付宝"`
 	Title            string     `json:"title" name:"购买的商品标题"`
-	Amount           int        `json:"amount" name:"金额，单位分"`
+	Amount           uint64     `json:"amount" name:"金额，单位分"`
 	ExtendedUserId   string     `json:"extended_user_id" bson:"extended_user_id" name:"平台用户标识"`
 	PrePaymentNo     string     `json:"pre_payment_no" bson:"pre_payment_no" name:"预下单号"`
 	PaymentNo        string     `json:"payment_no" bson:"payment_no" name:"支付单号"`
@@ -29,7 +29,7 @@ func (p Payment) ToAssociated() *AssociatedPayment {
 
 type AssociatedPayment struct {
 	Platform  string     `json:"platform" name:"支付平台:微信/支付宝"`
-	Amount    int        `json:"amount" name:"金额，单位分"`
+	Amount    uint64        `json:"amount" name:"金额，单位分"`
 	PaymentNo string     `json:"payment_no" bson:"payment_no" name:"支付单号"`
 	CreatedAt time.Time  `json:"created_at" bson:"created_at"` // 创建时间
 	PaymentAt *time.Time `json:"payment_at" bson:"payment_at"` // 支付时间
